pgo: Implement per-cpu profiling.
The profiler hook's try-locking caused inconsistent profile data
and it was also slow.

Since we prefer not to use exclusive locks in the profiler hook
solve the problem by duplicating the profiler data set for each cpu:
Index the prf_object->pcpu array by the smp_processor_id()
in the profiler hook.

The prf_object->data is now a a read-only template that is copied for each
prf_object_pcpu instance. The orginal section ptr is still needed for index
calculation tough.

This way each cpu profiles the kernel + modules independently and
they don't conflict any more in the profiler hook. :-)

Note: Reading the profile data is restricted to CPU0 with this commit.
Following commit adds a way to export profile data from all CPUs.

Signed-off-by: Jarmo Tiitto <jarmo.tiitto@gmail.com>